Factor out GFA parser
